Move Result.kt to the only module that used it, and delete core:common
No need for that core module now that the hilt coroutine is in external library
