--- core/common/README.md
@@ -1,3 +1,0 @@
-# :core:common module
-## Dependency graph
-![Dependency graph](../../docs/images/graphs/dep_graph_core_common.svg)
