+++ feature/topic/src/main/kotlin/com/google/samples/apps/nowinandroid/feature/topic/Result.kt
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 The Android Open Source Project
+ * Copyright 2024 The Android Open Source Project
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -14,8 +14,11 @@
  * limitations under the License.
  */
 
-package com.google.samples.apps.nowinandroid.core.result
+package com.google.samples.apps.nowinandroid.feature.topic
 
+import com.google.samples.apps.nowinandroid.feature.topic.Result.Error
+import com.google.samples.apps.nowinandroid.feature.topic.Result.Loading
+import com.google.samples.apps.nowinandroid.feature.topic.Result.Success
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.catch
 import kotlinx.coroutines.flow.map
@@ -27,6 +30,6 @@ sealed interface Result<out T> {
     data object Loading : Result<Nothing>
 }
 
-fun <T> Flow<T>.asResult(): Flow<Result<T>> = map<T, Result<T>> { Result.Success(it) }
-    .onStart { emit(Result.Loading) }
-    .catch { emit(Result.Error(it)) }
+fun <T> Flow<T>.asResult(): Flow<Result<T>> = map<T, Result<T>> { Success(it) }
+    .onStart { emit(Loading) }
+    .catch { emit(Error(it)) }
