+++ feature/topic/src/test/kotlin/com/google/samples/apps/nowinandroid/feature/topic/ResultKtTest.kt
