+++ settings.gradle.kts
@@ -37,7 +37,6 @@ include(":app")
 include(":app-nia-catalog")
 include(":benchmarks")
 include(":core:analytics")
-include(":core:common")
 include(":core:data")
 include(":core:data-test")
 include(":core:database")
